# demo/message-center-demo/contracts/openapi/message-center-v1.yaml
openapi: 3.1.0
info:
  title: CRM for Logistics Message Center API
  version: 1.0.0
  description: Versioned HTTP contract for the message center frontend and backend.
  contact:
    name: CRM for Logistics Engineering
servers:
  - url: /
...
security:
  - sessionCookie: []
paths:
...
  /api/v1/wecom/authorization/callback:
    get:
      operationId: verifyWeComAuthorizationCallback
      description: Verify the official delegated-authorization callback URL and return the decrypted echostr as plain text. This request is sent by WeCom when saving the callback URL and does not create or change an installation.
      tags:
        - Channels
      security: []
      parameters:
        - $ref: '#/components/parameters/TraceRequestId'
        - name: msg_signature
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 128
        - name: timestamp
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 32
            pattern: '^[0-9]{1,32}$'
        - name: nonce
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 128
        - name: echostr
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 1048576
      responses:
        '200':
          description: Decrypted callback verification string
          content:
            text/plain:
              schema:
                type: string
        '4XX':
          $ref: '#/components/responses/ApiErrorResponse'
        default:
          $ref: '#/components/responses/ApiErrorResponse'
    post:
      operationId: receiveWeComAuthorizationCallback
      description: Verify and decrypt the official delegated-authorization suite callback, then synchronously accept suite_ticket or enqueue bounded installation changes. The exact XML body is limited to 1 MiB and credential material is never returned.
      tags:
        - Channels
      security: []
      parameters:
        - $ref: '#/components/parameters/TraceRequestId'
        - name: msg_signature
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 128
        - name: timestamp
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 32
            pattern: '^[0-9]{1,32}$'
        - name: nonce
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 128
      requestBody:
        required: true
        content:
          application/xml:
            schema:
              $ref: '#/components/schemas/WeComAuthorizationCallbackBody'
      responses:
        '200':
          description: Official callback acknowledgement
          content:
            text/plain:
              schema:
                type: string
                const: success
        '4XX':
          $ref: '#/components/responses/ApiErrorResponse'
        default:
          $ref: '#/components/responses/ApiErrorResponse'
  /api/v1/wecom/app-callback:
    get:
      operationId: verifyWeComAppCallback
      description: Verify the application-level WeCom customer-contact callback URL and return the decrypted echostr. The endpoint is anonymous at the HTTP layer; the server validates the WeCom signature and application callback key.
      tags:
        - Channels
      security: []
      parameters:
        - $ref: '#/components/parameters/TraceRequestId'
        - name: msg_signature
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 128
        - name: timestamp
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 32
            pattern: '^[0-9]{1,32}$'
        - name: nonce
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 128
        - name: echostr
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 1048576
      responses:
        '200':
          description: Decrypted callback verification string
          content:
            text/plain:
              schema:
                type: string
        '403':
          $ref: '#/components/responses/ApiErrorResponse'
        '503':
          $ref: '#/components/responses/ApiErrorResponse'
        '4XX':
          $ref: '#/components/responses/ApiErrorResponse'
        default:
          $ref: '#/components/responses/ApiErrorResponse'
    post:
      operationId: receiveWeComAppCallback
      description: Verify and decrypt an application-level WeCom customer-contact event, then durably insert it before acknowledging success. The endpoint is anonymous at the HTTP layer; invalid signatures return 403 and unavailable persistence returns 503 so WeCom can retry. The request body is streamed into a bounded 1 MiB buffer; larger bodies return 413 before XML decoding.
      tags:
        - Channels
      security: []
      parameters:
        - $ref: '#/components/parameters/TraceRequestId'
        - name: msg_signature
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 128
        - name: timestamp
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 32
            pattern: '^[0-9]{1,32}$'
        - name: nonce
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 128
      requestBody:
        required: true
        content:
          application/xml:
            schema:
              $ref: '#/components/schemas/WeComAuthorizationCallbackBody'
          text/plain:
            schema:
              type: string
              minLength: 1
              maxLength: 1048576
      responses:
        '200':
          description: Event durably accepted or already known
          content:
            text/plain:
              schema:
                type: string
                const: success
        '403':
          $ref: '#/components/responses/ApiErrorResponse'
        '503':
          $ref: '#/components/responses/ApiErrorResponse'
        '413':
          $ref: '#/components/responses/ApiErrorResponse'
        '4XX':
          $ref: '#/components/responses/ApiErrorResponse'
        default:
          $ref: '#/components/responses/ApiErrorResponse'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: MESSAGE_CENTER_SESSION
      description: Opaque session token. The server stores only its SHA-256 digest.
    webhookSignature:
      type: apiKey
      in: header
      name: X-Webhook-Signature
      description: Header signature scheme. The adapter also requires X-Webhook-Timestamp and X-Webhook-Nonce for applicable channel types and verifies the exact raw body before parsing.
    weComSignature:
      type: apiKey
      in: query
      name: msg_signature
      description: WeCom query signature scheme. The adapter also requires timestamp and nonce query parameters and verifies the exact raw XML body before parsing.
  parameters:
    TraceRequestId:
      name: X-Request-ID
      in: header
      required: false
      schema:
        type: string
        minLength: 1
        maxLength: 128
...
  responses:
    ApiErrorResponse:
      description: Structured failure with no credential or signed URL material
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
...
    WeComAuthorizationCallbackBody:
      type: string
      format: binary
      minLength: 1
      maxLength: 1048576
    ApiError:
      type: object
      additionalProperties: false
      required: [code, message, traceId, fieldErrors]
      properties:
        code:
          type: string
          minLength: 1
          maxLength: 100
        message:
          type: string
          minLength: 1
          maxLength: 1000
        traceId:
          type: string
          minLength: 1
          maxLength: 128
        fieldErrors:
          type: array
          maxItems: 100
          items:
            $ref: '#/components/schemas/FieldError'
    FieldError:
      type: object
      additionalProperties: false
      required: [field, code, message]
      properties:
        field:
          type: string
          minLength: 1
          maxLength: 200
        code:
          type: string
          minLength: 1
          maxLength: 100
        message:
          type: string
          minLength: 1
          maxLength: 500
...